Add first working version of img override
This version transforms ![alt](src) in MDX into an Img component from
gatsby-image with all its goodness.

Taking some code and inspiration from:
https://github.com/gatsbyjs/gatsby/issues/10482#issuecomment-485349612

Which seems the best implementation until static query arguments are a
thing.

The beauty is that the component is used to override img tags with MDXProvider,
but can also be used directly. Notice that index.mdx is not modified in this
commit but still renders as before (even though the Image component was
modified).

There's still a problem: Img inserts a div that ends up inside a p. Will look
into fixing that next.