Update Dependencies for Node.js v13+
On Node 13, `yarn install` would fail on a `node-gyp` compilation
error. In order to get dependencies to install, you'd need to run `yarn
upgrade` to ensure packages are all up to date on the newest version. By
upgrading the dependencies in the distributed lockfile, it ensures that
future developers will be able to get the app started without too much
trouble.